Conflicts:
	Assets/Cenas/Final Scenes/SelectLanguage.unity
	Assets/Cenas/Final Scenes/WallBuilder.unity
	Assets/Cenas/New Scenes/MakeYourKitchen-NGUI CR 2.unity
	Assets/Custom Assets/GUI/Atlases/Introduction/Introduction.prefab
	Assets/Custom Assets/GUI/Atlases/Make Kitchen/Make Kitchen.png
	Assets/Custom Assets/GUI/Atlases/Make Kitchen/Make Kitchen.prefab
	Assets/Custom Assets/GUI/Fonts/Verdana/Verdana Bold.prefab

+++ Assets/Custom Assets/Scripts/GUI/Gadgets/GuiCatalogo.cs
@@ -388,17 +388,17 @@ public class GuiCatalogo : MonoBehaviour, GuiBase {
 									SomClique.Play();
 									selectColorDetalhe = i;
 							
-									Line.CurrentLine.GlobalDetailColorIndex = (uint)selectColorDetalhe;
+									Line.CurrentLine.GlobalDetailColorIndex = selectColorDetalhe;
 							
 									GameObject selectedMobile = GameObject.FindGameObjectWithTag("MovelSelecionado");
 									if(selectedMobile != null){
-										selectedMobile.GetComponent<InformacoesMovel>().ChangeDetailColor((uint)selectColorDetalhe);										
+										selectedMobile.GetComponent<InformacoesMovel>().ChangeDetailColor(selectColorDetalhe);										
 									}
 							
 									GameObject[] furniture = GameObject.FindGameObjectsWithTag("Movel");
 									if(furniture != null && furniture.Length != 0){
 										foreach(GameObject mobile in furniture){
-											mobile.GetComponent<InformacoesMovel>().ChangeDetailColor((uint)selectColorDetalhe);										
+											mobile.GetComponent<InformacoesMovel>().ChangeDetailColor(selectColorDetalhe);										
 										}
 									}
 								}
